wct第十一次

+++ src/main/java/com/btcd/controller/AdminController.java
@@ -1,5 +1,6 @@
 package com.btcd.controller;
 
+import com.baidu.ueditor.ActionEnter;
 import com.btcd.data.Admin;
 import com.btcd.data.Banner;
 import com.btcd.data.Bitclass;
@@ -7,22 +8,31 @@ import com.btcd.data.Project;
 import com.btcd.service.AdminService;
 import com.btcd.service.IndexService;
 import com.sun.org.apache.xpath.internal.operations.Mod;
+import org.apache.commons.fileupload.FileItemIterator;
+import org.apache.commons.fileupload.FileItemStream;
+import org.apache.commons.fileupload.FileUploadException;
+import org.apache.commons.fileupload.disk.DiskFileItemFactory;
+import org.apache.commons.fileupload.servlet.ServletFileUpload;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.util.DigestUtils;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
+import org.springframework.web.multipart.commons.CommonsMultipartResolver;
+
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.File;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.MalformedURLException;
 import java.sql.Date;
+import java.text.SimpleDateFormat;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.regex.Pattern;
 
 @Controller
@@ -31,6 +41,60 @@ public class AdminController {
     private AdminService adminService;
     @Autowired
     private IndexService indexService;
+
+    @RequestMapping(value = "/static/admin/ueditor/jsp/ueditor", method = RequestMethod.GET)
+    public @ResponseBody String ueditor(HttpServletRequest request, HttpServletResponse response) throws IllegalStateException, IOException {
+        //这里就是把controller.jsp代码copy下来
+        request.setCharacterEncoding( "utf-8" );
+        response.setHeader("Content-Type" , "text/html");
+        String roolPath = request.getSession().getServletContext().getRealPath("/");
+        String configStr = new ActionEnter(request, roolPath).exec();
+        System.out.printf("ueditor\n");
+        return configStr;
+    }
+
+    @RequestMapping(value = "/ueditorUpload", method = RequestMethod.POST)
+    public @ResponseBody Map<String, Object> ueUpload(HttpServletRequest request, HttpServletResponse response) throws IllegalStateException {
+        System.out.printf("uploading\n");
+        Map<String, Object> m = new HashMap<String, Object>();
+        CommonsMultipartResolver commonsMultipartResolver = new CommonsMultipartResolver(request.getSession().getServletContext());
+        commonsMultipartResolver.setDefaultEncoding("utf-8");
+        if(commonsMultipartResolver.isMultipart(request)){
+            MultipartHttpServletRequest multipartHttpServletRequest = (MultipartHttpServletRequest)request;
+            MultipartFile file = multipartHttpServletRequest.getFile("file");
+
+            if(!file.isEmpty()) {
+                //上传文件路径
+                String path = request.getServletContext().getRealPath("static"+File.separator+"admin"+File.separator+"ueditor"+File.separator+"upload");
+                //上传文件名
+                String filename = file.getOriginalFilename();
+                File filepath = new File(path,filename);
+                filepath.setWritable(true,false);
+                //判断路径是否存在，如果不存在就创建一个
+                if (!filepath.getParentFile().exists()) {
+                    filepath.getParentFile().mkdirs();
+                }
+                //将上传文件保存到一个目标文件当中
+                String newFileName = adminService.generateRandomFilename()+filename.substring(filename.lastIndexOf(".")+1);
+                File ultiPath = new File(path + File.separator + newFileName);
+                ultiPath.setWritable(true,false);
+                try {
+                    file.transferTo(ultiPath);
+                    m.put("path", "ueditor"+File.separator+"upload" + "/");
+                    m.put("filename", newFileName);
+                    m.put("original", filename);
+                    m.put("name", newFileName);
+                    m.put("url", "ueditor"+File.separator+"upload" + "/"+newFileName);
+                    m.put("state", "SUCCESS");
+                    m.put("type",filename.substring(filename.lastIndexOf(".")+1 ));
+                    m.put("size", file.getSize());
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+        return m;
+    }
 
     @RequestMapping("/admin")
     public String admin(HttpSession session){
@@ -201,7 +265,7 @@ public class AdminController {
     }
 
     @RequestMapping("/activityUpload")
-    public String activityUpload(Date endTime,boolean top, HttpServletRequest request, @RequestParam("image")MultipartFile file,HttpSession session){
+    public String activityUpload(Date endTime,boolean top, HttpServletRequest request,HttpSession session){
         if(session.getAttribute("admin" )==null){
             return "redirect:/adminLogin";
         }
@@ -216,31 +280,36 @@ public class AdminController {
         String address = request.getParameter("address");
         String state = request.getParameter("state");
         double price = Double.parseDouble(request.getParameter("price"));
+        String savePath = "img/portfolio/portfolio-6.jpg";
+        CommonsMultipartResolver commonsMultipartResolver = new CommonsMultipartResolver(request.getSession().getServletContext());
+        commonsMultipartResolver.setDefaultEncoding("utf-8");
         //如果文件不为空，写入上传路径
-        if(!file.isEmpty()) {
-            //上传文件路径
-            String path = request.getServletContext().getRealPath("static"+File.separator+"uploadFiles"+File.separator+"projects");
-            //上传文件名
-            String filename = file.getOriginalFilename();
-            File filepath = new File(path,filename);
-            filepath.setWritable(true,false);
-            //判断路径是否存在，如果不存在就创建一个
-            if (!filepath.getParentFile().exists()) {
-                filepath.getParentFile().mkdirs();
+        if(commonsMultipartResolver.isMultipart(request)){
+            MultipartHttpServletRequest multipartHttpServletRequest = (MultipartHttpServletRequest)request;
+            MultipartFile file = multipartHttpServletRequest.getFile("image");
+            if(!file.isEmpty()) {
+                //上传文件路径
+                String path = request.getServletContext().getRealPath("static"+File.separator+"uploadFiles"+File.separator+"projects");
+                //上传文件名
+                String filename = file.getOriginalFilename();
+                File filepath = new File(path,filename);
+                filepath.setWritable(true,false);
+                //判断路径是否存在，如果不存在就创建一个
+                if (!filepath.getParentFile().exists()) {
+                    filepath.getParentFile().mkdirs();
+                }
+                //将上传文件保存到一个目标文件当中
+                File ultiPath = new File(path + File.separator + adminService.generateRandomFilename()+filename.substring(filename.lastIndexOf(".")+1));
+                ultiPath.setWritable(true,false);
+                try {
+                    file.transferTo(ultiPath);
+                    savePath = ultiPath.getPath().substring(ultiPath.getPath().indexOf("uploadFiles"));
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
-            //将上传文件保存到一个目标文件当中
-            File ultiPath = new File(path + File.separator + adminService.generateRandomFilename()+filename.substring(filename.lastIndexOf(".")+1));
-            ultiPath.setWritable(true,false);
-            try {
-                file.transferTo(ultiPath);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            adminService.saveProject(title, ultiPath.getPath().substring(ultiPath.getPath().indexOf("uploadFiles")), content, method, address, state, price, endTime, top);
-        } else {
-            //TODO 文件为空时的情况（未完成）
-            return "redirect:/activityManage";
         }
+        adminService.saveProject(title,savePath, content, method, address, state, price, endTime, top);
         return "redirect:/activityManage";
     }
 
@@ -254,14 +323,16 @@ public class AdminController {
     }
 
     @RequestMapping("activityDelete/{id}")
-    public String activityDelete(@PathVariable("id")int id,HttpSession session){
+    public String activityDelete(@PathVariable("id")int id,HttpSession session,HttpServletRequest request){
         if(session.getAttribute("admin" )==null){
             return "redirect:/adminLogin";
         }
         Project project = adminService.findProjectById(id);
-        File file = new File(project.getPath());
-        if(file.exists()){
-            file.delete();
+        if(!project.getPath().equals("img/portfolio/portfolio-6.jpg")) {
+            File file = new File(request.getServletContext().getRealPath("static")+File.separator+project.getPath());
+            if (file.exists()) {
+                file.delete();
+            }
         }
         adminService.deleteProject(id);
         return "redirect:/activityManage";
@@ -277,7 +348,7 @@ public class AdminController {
     }
 
     @RequestMapping("activityModify/{id}")
-    public String activityModify(HttpSession session,@PathVariable("id")int id, Date endTime,boolean top, HttpServletRequest request, @RequestParam("image")MultipartFile file){
+    public String activityModify(HttpSession session,@PathVariable("id")int id, Date endTime,boolean top, HttpServletRequest request){
         if(session.getAttribute("admin" )==null){
             return "redirect:/adminLogin";
         }
@@ -292,7 +363,6 @@ public class AdminController {
         String address = request.getParameter("address");
         String state = request.getParameter("state");
         double price = Double.parseDouble(request.getParameter("price"));
-        System.out.printf("method:"+method);
         Project project = adminService.findProjectById(id);
         project.setTitle(title);
         project.setContent(content);
@@ -302,32 +372,40 @@ public class AdminController {
         project.setPrice(price);
         project.setEndTime(endTime);
         project.setTop(top);
+        CommonsMultipartResolver commonsMultipartResolver = new CommonsMultipartResolver(request.getSession().getServletContext());
+        commonsMultipartResolver.setDefaultEncoding("utf-8");
         //如果文件不为空，写入上传路径
-        if(!file.isEmpty()) {
-            //上传文件路径
-            String path = request.getServletContext().getRealPath("static"+File.separator+"uploadFiles"+File.separator+"projects");
-            //上传文件名
-            String filename = file.getOriginalFilename();
-            File filepath = new File(path,filename);
-            filepath.setWritable(true,false);
-            //判断路径是否存在，如果不存在就创建一个
-            if (!filepath.getParentFile().exists()) {
-                filepath.getParentFile().mkdirs();
+        if(commonsMultipartResolver.isMultipart(request)) {
+            MultipartHttpServletRequest multipartHttpServletRequest = (MultipartHttpServletRequest)request;
+            MultipartFile file = multipartHttpServletRequest.getFile("image");
+            if(!file.isEmpty()) {
+                //上传文件路径
+                String path = request.getServletContext().getRealPath("static"+File.separator+"uploadFiles"+File.separator+"projects");
+                //上传文件名
+                String filename = file.getOriginalFilename();
+                File filepath = new File(path,filename);
+                filepath.setWritable(true,false);
+                //判断路径是否存在，如果不存在就创建一个
+                if (!filepath.getParentFile().exists()) {
+                    filepath.getParentFile().mkdirs();
+                }
+                //将上传文件保存到一个目标文件当中
+                File ultiPath = new File(path + File.separator + adminService.generateRandomFilename()+filename.substring(filename.lastIndexOf(".")+1));
+                ultiPath.setWritable(true,false);
+                try {
+                    file.transferTo(ultiPath);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+                //删除之前的图片
+                if(!project.getPath().equals("img/portfolio/portfolio-6.jpg")) {
+                    File deleteFile = new File(request.getServletContext().getRealPath("static")+File.separator+project.getPath());
+                    if (deleteFile.exists()) {
+                        deleteFile.delete();
+                    }
+                }
+                project.setPath(ultiPath.getPath().substring(ultiPath.getPath().indexOf("uploadFiles")));
             }
-            //将上传文件保存到一个目标文件当中
-            File ultiPath = new File(path + File.separator + adminService.generateRandomFilename()+filename.substring(filename.lastIndexOf(".")+1));
-            ultiPath.setWritable(true,false);
-            try {
-                file.transferTo(ultiPath);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            //删除之前的图片
-            File deleteFile = new File(project.getPath());
-            if(deleteFile.exists()){
-                deleteFile.delete();
-            }
-            project.setPath(ultiPath.getPath().substring(ultiPath.getPath().indexOf("uploadFiles")));
         }
         adminService.updateProject(project);
         return "redirect:/activityManage";
